add helpers for generate element
