+++ lib/store/helpers/id/index.ts
@@ -1,0 +1,3 @@
+import { generateElementId } from "./generateElementId";
+import { getHighestElementId } from "./getHighestElementId";
+export { generateElementId, getHighestElementId };
